server/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
httpx==0.27.2
pydantic>=2.10,<3
python-dotenv==1.0.1
requests==2.32.3